moved current analysis modules to own file
