+++ src/test/integration/scheduler/test_regression_virtual_file_path.py
@@ -6,7 +6,7 @@ import pytest
 
 from intercom.back_end_binding import InterComBackEndBinding
 from objects.firmware import Firmware
-from scheduler.Analysis import AnalysisScheduler
+from scheduler.analysis import AnalysisScheduler
 from scheduler.Unpacking import UnpackingScheduler
 from storage.db_interface_backend import BackEndDbInterface
 from storage.MongoMgr import MongoMgr
