+++ src/test/integration/scheduler/test_unpack_analyse_and_compare.py
@@ -6,7 +6,7 @@ from unittest import TestCase, mock
 from helperFunctions.data_conversion import normalize_compare_id
 from helperFunctions.database import ConnectTo
 from objects.firmware import Firmware
-from scheduler.Analysis import AnalysisScheduler
+from scheduler.analysis import AnalysisScheduler
 from scheduler.Compare import CompareScheduler
 from scheduler.Unpacking import UnpackingScheduler
 from storage.db_interface_backend import BackEndDbInterface
