+++ src/test/unit/scheduler/test_analysis.py
@@ -7,7 +7,7 @@ from unittest import TestCase, mock
 import pytest
 
 from objects.firmware import Firmware
-from scheduler.Analysis import MANDATORY_PLUGINS, AnalysisScheduler
+from scheduler.analysis import MANDATORY_PLUGINS, AnalysisScheduler
 from test.common_helper import DatabaseMock, MockFileObject, fake_exit, get_config_for_testing, get_test_data_dir
 from test.mock import mock_patch, mock_spy
 
@@ -132,7 +132,7 @@ class TestAnalysisSchedulerBlacklist:
 
     @classmethod
     def setup_class(cls):
-        cls.init_patch = mock.patch(target='scheduler.Analysis.AnalysisScheduler.__init__', new=lambda *_: None)
+        cls.init_patch = mock.patch(target='scheduler.analysis.AnalysisScheduler.__init__', new=lambda *_: None)
         cls.init_patch.start()
         cls.sched = AnalysisScheduler()
         cls.sched.analysis_plugins = {}
@@ -229,7 +229,7 @@ class UtilityBase:
 
     @classmethod
     def setup_class(cls):
-        cls.init_patch = mock.patch(target='scheduler.Analysis.AnalysisScheduler.__init__', new=lambda *_: None)
+        cls.init_patch = mock.patch(target='scheduler.analysis.AnalysisScheduler.__init__', new=lambda *_: None)
         cls.init_patch.start()
         cls.scheduler = AnalysisScheduler()
         cls.scheduler.currently_running_lock = Manager().Lock()  # pylint: disable=no-member
@@ -356,7 +356,7 @@ class TestAnalysisSkipping:
 
     @classmethod
     def setup_class(cls):
-        cls.init_patch = mock.patch(target='scheduler.Analysis.AnalysisScheduler.__init__', new=lambda *_: None)
+        cls.init_patch = mock.patch(target='scheduler.analysis.AnalysisScheduler.__init__', new=lambda *_: None)
         cls.init_patch.start()
 
         cls.scheduler = AnalysisScheduler()
@@ -418,7 +418,7 @@ class TestAnalysisShouldReanalyse:
 
     @classmethod
     def setup_class(cls):
-        cls.init_patch = mock.patch(target='scheduler.Analysis.AnalysisScheduler.__init__', new=lambda *_: None)
+        cls.init_patch = mock.patch(target='scheduler.analysis.AnalysisScheduler.__init__', new=lambda *_: None)
         cls.init_patch.start()
         cls.scheduler = AnalysisScheduler()
         cls.init_patch.stop()
@@ -432,12 +432,12 @@ class TestAnalysisShouldReanalyse:
             # pylint: disable=unused-argument
             if plugin_name == 'plugin_root':
                 return plugin_root_date
-            elif plugin_name == 'plugin_dep':
+            if plugin_name == 'plugin_dep':
                 return plugin_dep_date
 
             assert False
 
-        monkeypatch.setattr('scheduler.Analysis._get_analysis_date', _get_analysis_date_mock)
+        monkeypatch.setattr('scheduler.analysis._get_analysis_date', _get_analysis_date_mock)
         uid = 'DONT_CARE'
         analysis_db_entry = {'plugin_version': '1.0'}
         plugin_mock = self.PluginMock()
